+++ plugin/app/Services/DefenceRegistrationService.php
@@ -5,6 +5,7 @@ namespace TTU\Charon\Services;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Log;
 use TTU\Charon\Exceptions\RegistrationException;
+use TTU\Charon\Models\Charon;
 use TTU\Charon\Models\Lab;
 use TTU\Charon\Repositories\CharonDefenseLabRepository;
 use TTU\Charon\Repositories\CharonRepository;
@@ -46,6 +47,8 @@ class DefenceRegistrationService
     private $submissionRepository;
 
     /**
+     * DefenceRegistrationService constructor.
+     *
      * @param CharonRepository $charonRepository
      * @param LabTeacherRepository $teacherRepository
      * @param DefenseRegistrationRepository $defenseRegistrationRepository
@@ -183,7 +186,7 @@ class DefenceRegistrationService
             }
         }
 
-        return array_keys(array_filter($labTimeslots, function($teachersRemaining) {
+        return array_keys(array_filter($labTimeslots, function ($teachersRemaining) {
             return $teachersRemaining < 1;
         }));
     }
@@ -199,26 +202,10 @@ class DefenceRegistrationService
      */
     public function validateRegistration(int $studentId, int $charonId, Lab $lab)
     {
-        $charonDuration = $this->charonRepository
-            ->getCharonById($charonId)
-            ->defense_duration;
+        $charon = $this->charonRepository->getCharonById($charonId);
 
-        if ($charonDuration == null || $charonDuration <= 0) {
+        if ($charon->defense_duration == null || $charon->defense_duration <= 0) {
             throw new RegistrationException('invalid_setup');
-        }
-
-        $registrations = $this->defenseRegistrationRepository->getDefenseRegistrationDurationsByLab($lab->id);
-        $totalOfDefenses = 0;
-        $labDurationInterval = $lab->start->diff($lab->end);
-        $labDuration = $labDurationInterval->h * 60 + $labDurationInterval->i;
-
-        $teacherCount = $this->teacherRepository->countLabTeachers($lab->id);
-        foreach ($registrations as $registration) {
-            $totalOfDefenses += $registration->defense_duration;
-        }
-
-        if ($labDuration * $teacherCount < $totalOfDefenses + $charonDuration) {
-            throw new RegistrationException("queue_full");
         }
 
         $pendingStudentDefences = $this->defenseRegistrationRepository->getUserPendingRegistrationsCount(
@@ -229,6 +216,10 @@ class DefenceRegistrationService
 
         if ($pendingStudentDefences > 0) {
             throw new RegistrationException('user_in_db');
+        }
+
+        if ($this->getEstimateTimeForNewRegistration($lab, $charon) === null) {
+            throw new RegistrationException("not_enough_time");
         }
     }
 
@@ -382,5 +373,77 @@ class DefenceRegistrationService
         );
 
         return 'inserted';
+    }
+
+    /**
+     * Give registrations their approximate starting time and sort them by it.
+     *
+     * @param array $registrations
+     * @param int $teacherCount
+     * @param Carbon $labStart
+     *
+     * @return array
+     */
+    public function attachEstimatedTimesToDefenceRegistrations(
+        array $registrations,
+        int $teacherCount,
+        Carbon $labStart
+    ): array {
+        $queuePresumption = array_fill(0, $teacherCount, 0);
+
+        for ($i = 0; $i < count($registrations); $i++) {
+            $teacherNr = array_keys($queuePresumption, min($queuePresumption))[0];
+
+            $registrations[$i]->estimated_start =
+                $labStart->copy()->addMinutes($queuePresumption[$teacherNr]);
+
+            $queuePresumption[$teacherNr] += $registrations[$i]->defense_duration;
+        }
+
+        return $registrations;
+    }
+
+    /**
+     * Calculate estimated starting time for a new defence registration.
+     *
+     * @param Lab $lab
+     * @param Charon $charon
+     *
+     * @return Carbon|null
+     */
+    public function getEstimateTimeForNewRegistration(Lab $lab, Charon $charon): ?Carbon
+    {
+        $capacity = $lab->end->diffInMinutes($lab->start);
+        $teacherCount = $this->teacherRepository->countLabTeachers($lab->id);
+
+        $registrations = $this->attachEstimatedTimesToDefenceRegistrations(
+            $this->defenseRegistrationRepository->getLabRegistrationsByLabId($lab->id, ['Waiting', 'Defending']),
+            $teacherCount,
+            $lab->start
+        );
+
+        if (count($registrations) >= $teacherCount) {
+
+            $latestRegistrations = array_slice($registrations, count($registrations) - $teacherCount);
+
+            $shortestWaitingTimeRegistration = array_reduce(
+                $latestRegistrations,
+                function ($r1, $r2) {
+                    return $r1 !== null && $r1->estimated_start < $r2->estimated_start
+                        ? $r1
+                        : $r2;
+                }
+            );
+
+            $shortestWaitingTime = $shortestWaitingTimeRegistration->estimated_start
+                ->addMinutes($shortestWaitingTimeRegistration->defense_duration);
+
+        } else {
+            $shortestWaitingTime = $lab->start;
+        }
+
+        return $capacity >= $shortestWaitingTime->diff($lab->start)->i + $charon->defense_duration
+            ? $shortestWaitingTime
+            : null;
     }
 }